+++ Source/Lib/TraktApiSharp/Objects/Get/Movies/TraktMovie.cs
@@ -62,7 +62,7 @@
         }
 
         [JsonProperty(PropertyName = "rating")]
-        public decimal? Rating { get; set; }
+        public float? Rating { get; set; }
 
         [JsonProperty(PropertyName = "votes")]
         public int? Votes { get; set; }
