+++ Source/Lib/TraktApiSharp/Objects/Get/Shows/Episodes/TraktEpisode.cs
@@ -66,7 +66,7 @@
         /// The average user rating of the episode.
         /// </summary>
         [JsonProperty(PropertyName = "rating")]
-        public decimal? Rating { get; set; }
+        public float? Rating { get; set; }
 
         /// <summary>
         /// The number of votes for the episode.
