+++ Source/Lib/TraktApiSharp/Objects/Get/Shows/Seasons/TraktSeason.cs
@@ -42,7 +42,7 @@
         /// The average user rating of the season.
         /// </summary>
         [JsonProperty(PropertyName = "rating")]
-        public decimal? Rating { get; set; }
+        public float? Rating { get; set; }
 
         /// <summary>
         /// The number of votes for the season.
